Isolate active x64 graph from D3DX8
Removes D3DX8 dependencies from active Evolution/WW3D and game-client sources by switching Bezier/point/sorting math to WWMath and replacing texture/surface helper usage with existing CPU BitmapHandler paths. Also fixes remaining native-width issues (pointer arithmetic/casts, alignment checks, buffer-size comparisons), updates related rendering code paths, and extends modernization docs and CMake policy tests to prevent D3DX8 and 32-bit pointer-truncation regressions in the x64 migration queue.

+++ Core/Libraries/Source/debug/debug_stack.cpp
@@ -33,6 +33,7 @@
 #include "WWLib/stringex.h"
 #include <imagehlp.h>
 #include <cinttypes>
+#include <cstddef>
 #include <cstdint>
 #include <cstdio>
 #include <cstring>
@@ -197,7 +198,7 @@ void DebugStackwalk::Signature::GetSymbol(std::uintptr_t addr, char *buf, unsign
   if (!gDbg._SymGetSymFromAddr(GetCurrentProcess(),static_cast<DWORD>(addr),&displacement,symPtr))
     return;
 #endif
-  if (static_cast<unsigned>(bufEnd-buf)<strlen(symPtr->Name)+24)
+  if (static_cast<std::size_t>(bufEnd-buf)<strlen(symPtr->Name)+24)
     return;
   buf+=sprintf(buf,", %s+0x%llx",symPtr->Name,static_cast<unsigned long long>(displacement));
 
@@ -218,7 +219,7 @@ void DebugStackwalk::Signature::GetSymbol(std::uintptr_t addr, char *buf, unsign
 
   p=strrchr(line.FileName,'\\');
   p=p?p+1:line.FileName;
-  if (static_cast<unsigned>(bufEnd-buf)<strlen(p)+24)
+  if (static_cast<std::size_t>(bufEnd-buf)<strlen(p)+24)
     return;
   sprintf(buf,", %s:%lu+0x%lx",p,static_cast<unsigned long>(line.LineNumber),static_cast<unsigned long>(lineDisplacement));
 }
